+++ Mage/FormPickerViewController.swift
@@ -7,63 +7,98 @@
 //
 import Foundation
 import MaterialComponents.MDCButton;
+import UIKit
 
 @objc protocol FormPickedDelegate {
     @objc func formPicked (form: Form);
     @objc func cancelSelection();
 }
 
+class ButtonFooterView: UICollectionReusableView {
+    var buttonDidTappedCallback: (() -> Void)?
+    var title: String? {
+        get {
+            return button.title(for: .normal)
+        }
+        set {
+            button.setTitle(newValue, for: .normal)
+            button.accessibilityLabel = newValue
+        }
+    }
+    
+    var _scheme: MDCContainerScheming?
+    var scheme: MDCContainerScheming? {
+        get {
+            return _scheme
+        }
+        set {
+            _scheme = newValue
+            if let scheme = newValue {
+                button.applyTextTheme(withScheme: scheme)
+                button.setTitleColor(scheme.colorScheme.primaryColorVariant, for: .normal)
+            }
+        }
+    }
+    
+    lazy var button: MDCButton = {
+        let button = MDCButton(forAutoLayout: ());
+        button.clipsToBounds = true;
+        button.addAction(UIAction(handler: { [unowned self] _ in
+            self.buttonDidTappedCallback?()
+        }), for: .touchUpInside)
+        return button;
+    }()
+    
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        configure()
+    }
+    
+    required init?(coder: NSCoder) {
+        fatalError("Not implemented")
+    }
+    
+    func configure() {
+        addSubview(button);
+        button.autoAlignAxis(toSuperviewAxis: .vertical);
+        button.autoPinEdge(toSuperviewEdge: .top, withInset: 12);
+        button.autoPinEdge(toSuperviewEdge: .bottom, withInset: 40);
+    }
+}
+
 @objc class FormPickerViewController: UIViewController {
+    
+    enum Section {
+        case main
+    }
 
     weak var delegate: FormPickedDelegate?;
-    var forms: [Form]?;
+    var forms: [Form] = []
     var scheme: MDCContainerScheming?;
     weak var observation: Observation?;
     var formIdCount: [Int : Int] = [ : ];
+    var didSetupConstraints = false
     
-    private lazy var divider: UIView = {
-        let divider = UIView(forAutoLayout: ());
-        divider.autoSetDimension(.height, toSize: 1);
-        return divider;
-    }()
+    var dataSource: UICollectionViewDiffableDataSource<Section, Form>?
+    var snapshot: NSDiffableDataSourceSnapshot<Section, Form>!
+    var cellRegistration: UICollectionView.CellRegistration<UICollectionViewListCell, Form>?
+    var footerRegistration: UICollectionView.SupplementaryRegistration<ButtonFooterView>?
+    var headerRegistration: UICollectionView.SupplementaryRegistration<UICollectionViewListCell>?
     
-    private lazy var titleLabel: UILabel = {
-        let titleLabel = UILabel(forAutoLayout: ());
-        titleLabel.text = "Add A Form To Your Observation";
-        return titleLabel;
-    }()
-    
-    private lazy var tableSectionHeaderView: UIView = {
-        let headerView = UIView(forAutoLayout: ());
-        headerView.addSubview(titleLabel)
-        headerView.addSubview(divider)
-        titleLabel.autoPinEdgesToSuperviewEdges(with: UIEdgeInsets(top: 0, left: 16, bottom: 0, right: 16))
-        divider.autoPinEdgesToSuperviewEdges(with: UIEdgeInsets.zero, excludingEdge: .top)
-        headerView.autoSetDimension(.height, toSize: 50)
-        headerView.autoSetDimension(.width, toSize: max(UIScreen.main.bounds.height, UIScreen.main.bounds.width))
-        return headerView;
-    }()
-    
-    lazy var tableView: UITableView = {
-        let tableView = UITableView(forAutoLayout: ());
-        tableView.accessibilityLabel = "Add A Form Table";
-        tableView.estimatedRowHeight = 100;
-        tableView.rowHeight = UITableView.automaticDimension;
-        tableView.insetsContentViewsToSafeArea = false;
-        return tableView;
-    }()
-    
-    private lazy var cancelButton: MDCButton = {
-        let cancelButton = MDCButton(forAutoLayout: ());
-        cancelButton.accessibilityLabel = "Cancel";
-        cancelButton.setTitle("Cancel", for: .normal);
-        cancelButton.clipsToBounds = true;
-        cancelButton.addTarget(self, action: #selector(cancelButtonTapped(_:)), for: .touchUpInside);
-        return cancelButton;
+    lazy var collectionView: UICollectionView = {
+        var configuration = UICollectionLayoutListConfiguration(appearance: .grouped)
+        configuration.headerMode = .supplementary
+        configuration.footerMode = .supplementary
+        
+        let collectionView = UICollectionView(frame: .zero, collectionViewLayout:  UICollectionViewCompositionalLayout.list(using: configuration))
+        collectionView.isAccessibilityElement = true
+        collectionView.accessibilityIdentifier = "Add A Form Table"
+        collectionView.accessibilityLabel = "Add A Form Table"
+        return collectionView
     }()
     
     @objc func cancelButtonTapped(_ sender: UIButton) {
-        delegate?.cancelSelection();
+        delegate?.cancelSelection()
     }
     
     func applyTheme(withScheme scheme: MDCContainerScheming? = nil) {
@@ -71,19 +106,13 @@ import MaterialComponents.MDCButton;
             return
         }
 
-        self.scheme = scheme;
+        self.scheme = scheme
         self.view.backgroundColor = scheme.colorScheme.backgroundColor
-        self.tableView.backgroundColor = scheme.colorScheme.backgroundColor;
-        cancelButton.applyTextTheme(withScheme: scheme);
-        cancelButton.setTitleColor(scheme.colorScheme.primaryColorVariant, for: .normal)
-        titleLabel.font = scheme.typographyScheme.body1
-        titleLabel.backgroundColor = scheme.colorScheme.surfaceColor
-        tableSectionHeaderView.backgroundColor = scheme.colorScheme.surfaceColor
-        divider.backgroundColor = scheme.colorScheme.onSurfaceColor.withAlphaComponent(0.12);
+        self.collectionView.backgroundColor = scheme.colorScheme.backgroundColor
     }
     
     init(frame: CGRect) {
-        super.init(nibName: nil, bundle: nil);
+        super.init(nibName: nil, bundle: nil)
     }
     
     required init(coder aDecoder: NSCoder) {
@@ -91,143 +120,162 @@ import MaterialComponents.MDCButton;
     }
     
     @objc public convenience init(delegate: FormPickedDelegate? = nil, forms: [Form]? = nil, observation: Observation? = nil, scheme: MDCContainerScheming?) {
-        self.init(frame: CGRect.zero);
-        self.delegate = delegate;
-        self.forms = forms;
-        self.observation = observation;
-        self.scheme = scheme;
-        applyTheme(withScheme: scheme);
+        self.init(frame: CGRect.zero)
+        self.delegate = delegate
+        if let forms = forms {
+            self.forms = forms
+        }
+        self.observation = observation
+        self.scheme = scheme
+        applyTheme(withScheme: scheme)
         if let observation = self.observation, let properties = observation.properties {
             if (properties.keys.contains(ObservationKey.forms.key)) {
-                let observationForms: [[String: Any]] = properties[ObservationKey.forms.key] as! [[String: Any]];
-                let formsToBeDeleted = observation.formsToBeDeleted;
+                let observationForms: [[String: Any]] = properties[ObservationKey.forms.key] as! [[String: Any]]
+                let formsToBeDeleted = observation.formsToBeDeleted
                 for (index, form) in observationForms.enumerated() {
                     if (!formsToBeDeleted.contains(index)) {
-                        let formId = form[EventKey.formId.key] as! Int;
-                        formIdCount[formId] = (formIdCount[formId] ?? 0) + 1;
+                        let formId = form[EventKey.formId.key] as! Int
+                        formIdCount[formId] = (formIdCount[formId] ?? 0) + 1
                     }
                 }
             }
         }
     }
     
+    func footerRegistrationHandler(footerView: ButtonFooterView, elementKind: String, indexPath: IndexPath) {
+        footerView.title = "Cancel"
+        footerView.buttonDidTappedCallback = { [weak self] in
+            self?.delegate?.cancelSelection()
+        }
+        if let scheme = self.scheme {
+            footerView.scheme = scheme
+        }
+    }
+    
+    func headerRegistrationHandler(headerView: UICollectionViewListCell, elementKind: String, indexPath: IndexPath) {
+        var configuration = headerView.defaultContentConfiguration()
+        var backgroundConfiguration = UIBackgroundConfiguration.listGroupedHeaderFooter()
+        
+        configuration.text = "Add A Form To Your Observation"
+        
+        if let scheme = self.scheme {
+            configuration.textProperties.font = scheme.typographyScheme.body1
+            backgroundConfiguration.backgroundColor = scheme.colorScheme.surfaceColor
+        }
+        
+        headerView.contentConfiguration = configuration
+        headerView.backgroundConfiguration = backgroundConfiguration
+    }
+    
+    func cellRegistrationHandler(cell: UICollectionViewListCell, indexPath: IndexPath, item: Form) {
+        var configuration = cell.defaultContentConfiguration()
+        configuration.text = item.name
+        configuration.secondaryText = item.formDescription
+        
+        configuration.image = UIImage(systemName: "doc.text.fill")?.aspectResize(to: CGSize(width: 40, height: 40)).withRenderingMode(.alwaysTemplate)
+        configuration.imageProperties.maximumSize = CGSize(width: 40, height: 40)
+        configuration.imageProperties.reservedLayoutSize = CGSize(width: 40, height: 40)
+        
+        let formCount = self.formIdCount[item.formId?.intValue ?? Int.min] ?? 0
+        let formMin: Int = item.min ?? 0
+        let formMax: Int = item.max ?? Int.max
+        
+        if (formCount < formMin) {
+            configuration.text = "\(configuration.text ?? "")*"
+        }
+        
+        var backgroundColor: UIColor = .clear
+        if (formCount >= formMax) {
+            configuration.imageProperties.tintColor = globalDisabledScheme().colorScheme.onSurfaceColor
+            configuration.textProperties.color = globalDisabledScheme().colorScheme.onSurfaceColor
+            configuration.secondaryTextProperties.color = globalDisabledScheme().colorScheme.onSurfaceColor
+            backgroundColor = globalDisabledScheme().colorScheme.surfaceColor
+            
+        } else {
+            if let color = item.color {
+                configuration.imageProperties.tintColor = UIColor(hex: color);
+            } else {
+                configuration.imageProperties.tintColor = self.scheme?.colorScheme.primaryColor
+            }
+            if let scheme = self.scheme {
+                configuration.textProperties.color = scheme.colorScheme.onSurfaceColor
+                configuration.textProperties.font = scheme.typographyScheme.subtitle1
+                configuration.secondaryTextProperties.color = scheme.colorScheme.onSurfaceColor.withAlphaComponent(0.6)
+                backgroundColor = scheme.colorScheme.surfaceColor
+            }
+        }
+        
+        var backgroundConfiguration = UIBackgroundConfiguration.listGroupedCell()
+        backgroundConfiguration.backgroundColor = backgroundColor
+        cell.contentConfiguration = configuration
+        cell.backgroundConfiguration = backgroundConfiguration
+    }
+    
     override func loadView() {
-        super.loadView();
-        view.addSubview(tableView);
+        super.loadView()
+        view.addSubview(collectionView)
+        collectionView.autoPinEdgesToSuperviewEdges(with: .zero, excludingEdge: .top)
+        collectionView.autoPinEdge(toSuperviewMargin: .top)
     }
     
     override func viewDidLoad() {
         super.viewDidLoad();
-        self.tableView.delegate = self;
-        self.tableView.dataSource = self;
-        tableView.autoPinEdgesToSuperviewEdges();
-        tableView.contentInsetAdjustmentBehavior = .never;
-    }
-}
-
-extension FormPickerViewController: UITableViewDataSource {
-
-    func numberOfSections(in tableView: UITableView) -> Int {
-        return 1;
-    }
-
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return self.forms?.count ?? 0;
-    }
-
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell: UITableViewCell = {
-            guard let cell = tableView.dequeueReusableCell(withIdentifier: "formCell") else {
-                // Never fails:
-                return UITableViewCell(style: .subtitle, reuseIdentifier: "formCell")
-            }
+        
+        footerRegistration = UICollectionView.SupplementaryRegistration
+        <ButtonFooterView>(elementKind: UICollectionView.elementKindSectionFooter, handler: footerRegistrationHandler)
+        
+        headerRegistration = UICollectionView.SupplementaryRegistration
+        <UICollectionViewListCell>(elementKind: UICollectionView.elementKindSectionHeader, handler: headerRegistrationHandler)
+        
+        cellRegistration = UICollectionView.CellRegistration<UICollectionViewListCell, Form>(handler: cellRegistrationHandler)
+        
+        dataSource = UICollectionViewDiffableDataSource<Section, Form>(collectionView: collectionView) {
+            (collectionView: UICollectionView, indexPath: IndexPath, identifier: Form) -> UICollectionViewCell? in
+            
+            let cell = collectionView.dequeueConfiguredReusableCell(
+                using: self.cellRegistration!, for: indexPath, item: identifier)
+            
             return cell
-        }()
-        cell.backgroundColor = scheme?.colorScheme.surfaceColor;
-
-        if let form = self.forms?[indexPath.row] {
-            cell.accessibilityLabel = form.name;
-            cell.textLabel?.text = form.name
-            cell.detailTextLabel?.text = form.formDescription;
-            cell.imageView?.image = UIImage(systemName: "doc.text.fill")?.aspectResize(to: CGSize(width: 40, height: 40)).withRenderingMode(.alwaysTemplate);
-            
-            
-            let formCount = formIdCount[form.formId?.intValue ?? Int.min] ?? 0;
-            let formMin: Int = form.min ?? 0;
-            let formMax: Int = form.max ?? Int.max;
-            
-            if (formCount < formMin) {
-                cell.textLabel?.text = "\(cell.textLabel?.text ?? "")*";
-            }
-            
-            if (formCount >= formMax) {
-                cell.imageView?.tintColor = globalDisabledScheme().colorScheme.onSurfaceColor
-                cell.textLabel?.textColor = globalDisabledScheme().colorScheme.onSurfaceColor
-                cell.detailTextLabel?.textColor = globalDisabledScheme().colorScheme.onSurfaceColor;
-                cell.backgroundColor = globalDisabledScheme().colorScheme.surfaceColor;
-            } else {
-                if let color = form.color {
-                    cell.imageView?.tintColor = UIColor(hex: color);
-                } else {
-                    cell.imageView?.tintColor = scheme?.colorScheme.primaryColor
-                }
-                cell.textLabel?.textColor = scheme?.colorScheme.onSurfaceColor;
-                cell.detailTextLabel?.textColor = scheme?.colorScheme.onSurfaceColor.withAlphaComponent(0.6);
-                cell.backgroundColor = scheme?.colorScheme.surfaceColor;
-                cell.textLabel?.font = scheme?.typographyScheme.subtitle1;
-            }
         }
-
-        return cell;
+        
+        dataSource?.supplementaryViewProvider = { [unowned self]
+            (collectionView, elementKind, indexPath) -> UICollectionReusableView? in
+            
+            if elementKind == UICollectionView.elementKindSectionHeader {
+                return collectionView.dequeueConfiguredReusableSupplementary(using: headerRegistration!, for: indexPath)
+            }
+            return collectionView.dequeueConfiguredReusableSupplementary(using: footerRegistration!, for: indexPath)
+        }
+        
+        snapshot = NSDiffableDataSourceSnapshot<Section, Form>()
+        snapshot.appendSections([.main])
+        snapshot.appendItems(forms, toSection: .main)
+        
+        dataSource?.apply(snapshot, animatingDifferences: false)
+        
+        self.collectionView.delegate = self
     }
 }
 
-extension FormPickerViewController: UITableViewDelegate {
-    func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-        if let form = self.forms?[indexPath.row] {
-            if (form.formDescription != nil) {
-                return 72.0
-            } else {
-                return 56.0
-            }
+extension FormPickerViewController : UICollectionViewDelegate {
+    func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
+        guard let form = dataSource?.itemIdentifier(for: indexPath) else {
+            collectionView.deselectItem(at: indexPath, animated: true)
+            return
         }
-        return UITableView.automaticDimension;
-    }
-
-    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        if let form = self.forms?[indexPath.row] {
-            
-            let formCount = formIdCount[form.formId?.intValue ?? Int.min] ?? 0;
-            let formMax: Int = form.max ?? Int.max;
-            
-            if (formCount >= formMax) {
-                // max amount of this form have already been added
-                let message: MDCSnackbarMessage = MDCSnackbarMessage(text: "\(form.name ?? "") form cannot be included in an observation more than \(formMax) time\(formMax == 1 ? "" : "s")");
-                let messageAction = MDCSnackbarMessageAction();
-                messageAction.title = "OK";
-                message.action = messageAction;
-                MDCSnackbarManager.default.show(message);
-            } else {
-                delegate?.formPicked(form: form);
-            }
+        
+        let formCount = formIdCount[form.formId?.intValue ?? Int.min] ?? 0
+        let formMax: Int = form.max ?? Int.max
+        
+        if (formCount >= formMax) {
+            // max amount of this form have already been added
+            let message: MDCSnackbarMessage = MDCSnackbarMessage(text: "\(form.name ?? "") form cannot be included in an observation more than \(formMax) time\(formMax == 1 ? "" : "s")")
+            let messageAction = MDCSnackbarMessageAction()
+            messageAction.title = "OK"
+            message.action = messageAction
+            MDCSnackbarManager.default.show(message)
+        } else {
+            delegate?.formPicked(form: form)
         }
-    }
-
-    func tableView(_ tableView: UITableView, viewForFooterInSection section: Int) -> UIView? {
-        let footerView = UIView(frame: .zero);
-        footerView.addSubview(cancelButton);
-        cancelButton.autoAlignAxis(toSuperviewAxis: .vertical);
-        cancelButton.autoPinEdge(toSuperviewEdge: .top, withInset: 12);
-        cancelButton.autoPinEdge(toSuperviewEdge: .bottom, withInset: 20);
-        footerView.backgroundColor = scheme?.colorScheme.backgroundColor
-        return footerView;
-    }
-    
-    func tableView(_ tableView: UITableView, viewForHeaderInSection section: Int) -> UIView? {
-        return tableSectionHeaderView
-    }
-    
-    func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
-        return 50
     }
 }