+++ Mage/MAGE-Bridging-Header.h
@@ -59,7 +59,6 @@
 #import "ObservationTableHeaderView.h"
 #import "StaticPointAnnotation.h"
 #import "LocationFilterTableViewController.h"
-#import "EventTableViewCell.h"
 
 #import "AttachmentPushService.h"
 #import "LocationAnnotation.h"